requirements.txt.txt
selenium
pytest
allure-pytest
flake8